requirements.txt
flask~=3.0.0
requests~=2.31.0
Pillow~=10.0.1
ttkthemes~=3.2.2
matplotlib~=3.8.0
seaborn~=0.13.0
numpy~=1.26.1
pandas~=2.1.2
tkcalendar~=1.6.1